adding tts for voice and layaring functions
