+++ requirements.txt
@@ -8,3 +8,5 @@ mysql-connector-python~=9.2.0
 python-dotenv~=1.0.1
 tkcalendar~=1.6.1
 requests~=2.32.3
+pygame
+gTTs
